Split widget: computed row
